nlp-requirements.txt
torch
torchaudio
numpy
fastapi
uvicorn
llama-cpp-python
